fix(refactoring): create a group to game view
