+++ spelling-bubbles/Game/ContentView.swift
@@ -9,21 +9,6 @@ import SwiftUI
 import SpriteKit
 
 
-struct GameView: UIViewControllerRepresentable {
-    
-    @EnvironmentObject private var viewManager: ViewManager
-    @EnvironmentObject private var levelManager: LevelManager
-    
-    typealias UIViewControllerType = GameViewController
-    
-    func makeUIViewController(context: Context) -> GameViewController {
-        return GameViewController(viewManager: viewManager,
-                                  levelManager: levelManager)
-    }
-    
-    func updateUIViewController(_ uiViewController: GameViewController, context: Context) {}
-}
-
 struct ContentView: View {
     
     @EnvironmentObject private var viewManager: ViewManager
